feat: update
1. fix cmake failed in protobuf on Linux
2. fix compile error
3. add sqlite backend in client side in order to debug

+++ controller/cli/queuelist.hpp
@@ -45,7 +45,7 @@ public:
 
     QueueList();
 
-    i32 init();
+    u8 init();
 
     i32 run();
